+++ WebServiceSmartKitchen/Models/Model.Context.cs
@@ -10,14 +10,41 @@
 namespace WebServiceSmartKitchen.Models
 {
     using System;
+    using System.Linq;
     using System.Data.Entity;
     using System.Data.Entity.Infrastructure;
+    using System.Data.Entity.Validation;
+    using System.Threading.Tasks;
     
     public partial class ModelContainer : DbContext
     {
         public ModelContainer()
             : base("name=ModelContainer")
         {
+        }
+
+        public override async Task<int> SaveChangesAsync()
+        {
+            try
+            {
+                return await base.SaveChangesAsync();
+            }
+            catch (DbEntityValidationException ex)
+            {
+                // Retrieve the error messages as a list of strings.
+                var errorMessages = ex.EntityValidationErrors
+                        .SelectMany(x => x.ValidationErrors)
+                        .Select(x => x.ErrorMessage);
+
+                // Join the list to a single string.
+                var fullErrorMessage = string.Join("; ", errorMessages);
+
+                // Combine the original exception message with the new one.
+                var exceptionMessage = string.Concat(ex.Message, " The validation errors are: ", fullErrorMessage);
+
+                // Throw a new DbEntityValidationException with the improved exception message.
+                throw new DbEntityValidationException(exceptionMessage, ex.EntityValidationErrors);
+            }
         }
     
         protected override void OnModelCreating(DbModelBuilder modelBuilder)
